优化一些代码

+++ src/main/java/cn/felord/domain/contactbook/department/DeptInfo.java
@@ -3,13 +3,12 @@ package cn.felord.domain.contactbook.department;
 import lombok.Data;
 
 /**
- *
  * @author n1
  */
 @Data
-public class DeptInfo{
-    private Integer id;
-    private Integer parentid;
+public class DeptInfo {
+    private Long id;
+    private Long parentid;
     private String name;
     private String nameEn;
     private Long order;
